+++ lib/nretnil/compendium-api/compendium.rb
@@ -11,6 +11,11 @@ module Nretnil
       if ENV['DEBUG'] == 'true'
         debug_output $stderr
       end
+      unless ENV['http_proxy'] == '' || ENV['http_proxy'].nil?
+        proxy = ENV['http_proxy'].gsub('http://','').gsub('https://','').split(':')
+        puts proxy
+        http_proxy proxy[0], proxy[1]
+      end
 
       def initialize(user, api_key, server)
         @auth = {:username => user, :password => api_key}